// src/interfaces/dashboard.interface.ts
export interface DashboardJson {
  all_userss: number;
  all_users_actives: number;
  all_users_inactives: number;
  all_transactions: number;
  all_trans_deposits: number;
  all_trans_withdrawals: number;
  all_trans_subscribs: number;
  all_trans_rewards: number;
  all_trans_disbursements: number;
  all_reclamations: number;
  all_reclamation_pendings: number;
  // all_vip_subscribtions: number;
  // all_subscribtions: number;
  sum_reward_amounts: number;
  all_advertisements: number;
  all_coupons: number;
  all_coupon_pendings: number;
  // all_coupon_wins: number;
  // all_coupon_loses: number;
  // all_events: number;
  // all_event_pendings: number;
  // all_event_wins: number;
  // all_event_loses: number;
  // all_matchs: number;
  // all_championnats: number;
  // all_sports: number;
  // all_clubs: number;
  all_bonus: number;
  all_service_transactions: number;
  // all_coobet_transactions: number;
  withdrawals_amount: number;
  deposits_amount: number;
  xbet_transaction: number;
  betwinner_transaction: number;
  starz_transaction: number;
  melbet_transaction: number;
  all_bot_trans_deposit: number;
  all_bot_trans_withdrawal: number;
  all_bot_transaction: number;
  all_telegram_user: number;
}